Add Wasmtime, in Libraries / Virtualization.
I'm open to ideas for putting this in a different category.

WebAssembly itself doesn't neatly fit within the existing categories.
I picked virtualization here because that's one of the things it's known
for. Similarly, Wasmtime can be used as a standalone application, but
it's more known for being used as a library.